[Compile][Platform] Make PiecewiseBackend pluggable and extendable
